+++ src/Services/ConfigReaderFilterDecorator/Job/JobFilterParametersInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Trivago\Rumi\Services\ConfigReaderFilterDecorator\Job;
+
+interface JobFilterParametersInterface
+{
+    public function getJobFilter(): string;
+}