+++ tests/Commands/RunCommandTest.php
@@ -42,6 +42,7 @@ use Trivago\Rumi\Models\RunConfig;
 use Trivago\Rumi\Models\StagesCollection;
 use Trivago\Rumi\Process\RunningProcessesFactory;
 use Trivago\Rumi\Services\ConfigReader;
+use Trivago\Rumi\Services\ConfigReaderInterface;
 
 /**
  * @covers \Trivago\Rumi\Commands\RunCommand
@@ -96,7 +97,7 @@ class RunCommandTest extends TestCase
         $this->processFactory = $this->prophesize(RunningProcessesFactory::class);
         $this->container->set('trivago.rumi.process.running_processes_factory', $this->processFactory->reveal());
 
-        $this->configReader = $this->prophesize(ConfigReader::class);
+        $this->configReader = $this->prophesize(ConfigReaderInterface::class);
         $this->command = new RunCommand(
             $this->eventDispatcher->reveal(),
             $this->configReader->reveal(),
@@ -109,7 +110,7 @@ class RunCommandTest extends TestCase
     public function testGivenNoCiYamlFile_WhenExecuted_ThenDisplaysErrorMessage()
     {
         // given
-        $this->configReader->getRunConfig(Argument::any(), Argument::is(CommandAbstract::DEFAULT_CONFIG))->willThrow(new \Exception(
+        $this->configReader->getRunConfig()->willThrow(new \Exception(
             'Required file \''.CommandAbstract::DEFAULT_CONFIG.'\' does not exist',
             ReturnCodes::RUMI_YML_DOES_NOT_EXIST
         ));
@@ -125,7 +126,7 @@ class RunCommandTest extends TestCase
     public function testGivenCiYamlSyntaxIsWrong_WhenExecuted_ThenDisplaysErrorMessage()
     {
         // given
-        $this->configReader->getRunConfig(Argument::any(), Argument::is(CommandAbstract::DEFAULT_CONFIG))->willThrow(new ParseException(
+        $this->configReader->getRunConfig()->willThrow(new ParseException(
             'Unable to parse at line 2 (near "::yaml_file").'
         ));
 
@@ -145,7 +146,7 @@ class RunCommandTest extends TestCase
             $this->getTearDownProcess()
         );
 
-        $this->configReader->getRunConfig(Argument::any(), Argument::is(CommandAbstract::DEFAULT_CONFIG))
+        $this->configReader->getRunConfig()
             ->willReturn(
                 new RunConfig(
                     new StagesCollection(
@@ -180,7 +181,7 @@ class RunCommandTest extends TestCase
 
         $this->setProcessFactoryMock($startProcess, $tearDownProcess);
 
-        $this->configReader->getRunConfig(Argument::any(), Argument::is(CommandAbstract::DEFAULT_CONFIG))
+        $this->configReader->getRunConfig()
             ->willReturn(
                 new RunConfig(
                     new StagesCollection(
@@ -219,7 +220,7 @@ class RunCommandTest extends TestCase
 
         $this->setProcessFactoryMock($startProcess, $tearDownProcess);
 
-        $this->configReader->getRunConfig(Argument::any(), Argument::is(CommandAbstract::DEFAULT_CONFIG))
+        $this->configReader->getRunConfig()
             ->willReturn(
                 new RunConfig(
                     new StagesCollection(
@@ -254,7 +255,7 @@ class RunCommandTest extends TestCase
 
         $this->setProcessFactoryMock($startProcess, $tearDownProcess);
 
-        $this->configReader->getRunConfig(Argument::any(), Argument::is(CommandAbstract::DEFAULT_CONFIG))
+        $this->configReader->getRunConfig()
             ->willReturn(
                 new RunConfig(
                     new StagesCollection(
@@ -329,7 +330,7 @@ class RunCommandTest extends TestCase
 
         $this->setProcessFactoryMock($startProcess, $tearDownProcess);
 
-        $this->configReader->getRunConfig(Argument::any(), Argument::is(CommandAbstract::DEFAULT_CONFIG))
+        $this->configReader->getRunConfig()
             ->willReturn(
                 new RunConfig(
                     new StagesCollection(
@@ -392,7 +393,7 @@ class RunCommandTest extends TestCase
         $input = new ArrayInput(['--config' => $configFile]);
 
         $this->configReader
-            ->getRunConfig(Argument::any(), Argument::is($configFile))
+            ->getRunConfig()
             ->willThrow(new \Exception($exceptionMessage, ReturnCodes::RUMI_YML_DOES_NOT_EXIST))
             ->shouldBeCalledTimes(1);